+++ neoforge/src/main/java/com/faboslav/friendsandfoes/modcompat/neoforge/CuriosCompat.java
@@ -3,7 +3,6 @@ package com.faboslav.friendsandfoes.modcompat.neoforge;
 import com.faboslav.friendsandfoes.modcompat.ModCompat;
 
 import java.util.EnumSet;
-/*
 import com.faboslav.friendsandfoes.events.lifecycle.ClientSetupEvent;
 import com.faboslav.friendsandfoes.init.FriendsAndFoesItems;
 import com.faboslav.friendsandfoes.modcompat.neoforge.curios.CuriosTotemRenderer;
@@ -11,21 +10,18 @@ import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
-import net.neoforged.fml.InterModComms;
 import org.jetbrains.annotations.Nullable;
 import top.theillusivec4.curios.api.CuriosApi;
 import top.theillusivec4.curios.api.SlotResult;
 import top.theillusivec4.curios.api.client.CuriosRendererRegistry;
 
 import java.util.function.Predicate;
- */
 
 @SuppressWarnings({"all", "removal"})
 public final class CuriosCompat implements ModCompat
 {
 	public CuriosCompat() {
-		//InterModComms.sendTo("curios", top.theillusivec4.curios.api.SlotTypeMessage.REGISTER_TYPE, () -> top.theillusivec4.curios.api.SlotTypePreset.CHARM.getMessageBuilder().build());
-		//ClientSetupEvent.EVENT.addListener(CuriosCompat::registerRenderers);
+		ClientSetupEvent.EVENT.addListener(CuriosCompat::registerRenderers);
 	}
 
 	@Override
@@ -33,7 +29,6 @@ public final class CuriosCompat implements ModCompat
 		return EnumSet.of(Type.CUSTOM_EQUIPMENT_SLOTS);
 	}
 
-	/*
 	private static void registerRenderers(final ClientSetupEvent clientSetupEvent) {
 		CuriosRendererRegistry.register(Items.TOTEM_OF_UNDYING, CuriosTotemRenderer::new);
 		CuriosRendererRegistry.register(FriendsAndFoesItems.TOTEM_OF_FREEZING.get(), CuriosTotemRenderer::new);
@@ -48,5 +43,5 @@ public final class CuriosCompat implements ModCompat
 		}
 
 		return null;
-	}*/
+	}
 }