+++ neoforge/src/main/java/com/faboslav/friendsandfoes/modcompat/neoforge/curios/CuriosTotemRenderer.java
@@ -1,6 +1,5 @@
 package com.faboslav.friendsandfoes.modcompat.neoforge.curios;
 
-/*
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.render.OverlayTexture;
 import net.minecraft.client.render.VertexConsumerProvider;
@@ -50,4 +49,4 @@ public class CuriosTotemRenderer implements ICurioRenderer
 			0
 		);
 	}
-}*/
+}