+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.17
+
+- Fixed tuff golem holding enchanted "book" instead of enchanted "enchanted book"
+- Fixed mauler being able to be infinitely fed using the same item
+- Fixed trinkets and curios compat
+
 ## 2.0.16
 
 - Fixed glares breeding